requirements.txt
# MoFrame dependencies

# Core
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Video
moviepy>=1.0.3
imageio-ffmpeg>=0.4.9

# Archives
rarfile>=4.0

# PDF
PyMuPDF>=1.23.0
pdf2image>=1.16.3

# UI
streamlit>=1.28.0

# Utils
tqdm>=4.66.0

# Testing
pytest>=7.4.0
